feat [PM4-60] refactor code
